editor/libeditor/html/tests/test_bug551704.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=551704
-->
<head>
  <title>Test for Bug 551704</title>
  <script type="application/javascript" src="/MochiKit/packed.js"></script>
  <script type="application/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <script type="text/javascript" src="/tests/SimpleTest/EventUtils.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=551704">Mozilla Bug 551704</a>
<p id="display"></p>
<div id="content">
  <div id="preformatted" style="white-space: pre" contenteditable>a&#10;b</div>
</div>
<pre id="test">
<script type="application/javascript">

/** Test for Bug 551704 **/
SimpleTest.waitForExplicitFinish();
addLoadEvent(function() {
  var preformatted = document.getElementById("preformatted");
  is(preformatted.innerHTML, "a\nb", "No BR node should be injected for preformatted editable fields");

  SimpleTest.finish();
});

</script>
</pre>
</body>
</html>
